requirements.txt
openai
sentence-transformers
faiss-cpu
pandas
matplotlib
pydantic
python-dotenv
fitz   # PyMuPDF
PyMuPDF==1.24.9
requests
tf-keras